// src/_common/exceptions/error-code.enum.ts
export enum ErrorCodeEnum {
  PERMISSION_DENIED = 403,
  UNAUTHORIZED = 600,
  EMAIL_ALREADY_EXISTS = 601,
  PHONE_ALREADY_EXISTS = 602,
  INVALID_CREDENTIALS = 603,
  BLOCKED_USER = 604,
  USER_PHONE_NOT_VERIFIED_YET = 605,
  USER_DOES_NOT_EXIST = 606,
  USER_EMAIL_IS_NOT_VERIFIED_YET = 607,
  VERIFICATION_CODE_NOT_EXIST = 608,
  EXPIRED_VERIFICATION_CODE = 609,
  INVALID_SESSION = 610,
  INVALID_REFRESH_TOKEN = 635,
  WRONG_PASSWORD = 611,
  CONFIGURATION_SHOULD_HAS_UNIQUE_KEY = 612,
  CONFIGURATION_NOT_EXISTS = 613,
  NOTIFICATION_DOES_NOT_EXIST = 614,
  USER_PHONE_ALREADY_VERIFIED = 617,
  INCORRECT_DATE = 619,
  FROM_VALUE_HIGHER_THAN_OR_EQUAL_TO_VALUE = 620,
  SECURITY_GROUP_NAME_ALREADY_EXISTS = 624,
  SECURITY_GROUP_DOES_NOT_EXIST = 625,
  INCORRECT_PHONE_OR_PASSWORD = 628,
  INCORRECT_EMAIL_OR_PASSWORD = 630,
  CONFIRM_PASSWORD_DOESN_T_MATCH = 634,
  INVALID_EMAIL = 641,
  INVALID_PHONE_NUMBER = 642,
  CONTACT_MESSAGE_NOT_EXIST = 681,
  STORED_PAYMENT_METHOD_NOT_EXIST = 689,
  CANT_DELETE_SUPER_ADMIN_GROUP = 692,
}