// src/user/user.enums.ts
export enum Lang {
  AR = 'AR',
  EN = 'EN',
}

export enum Gender {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
}

export enum UserRole {
  USER = 'USER',
  ADMIN = 'ADMIN',
}

export enum DeviceEnum {
  DESKTOP = 'DESKTOP',
  IOS = 'IOS',
  ANDROID = 'ANDROID',
}

export enum UserVerificationCodeUseCaseEnum {
  PASSWORD_RESET = 'PASSWORD_RESET',
  PHONE_VERIFICATION = 'PHONE_VERIFICATION',
  EMAIL_VERIFICATION = 'EMAIL_VERIFICATION',
}